use prop picker context as a shared context
